Only use `prev_nets` to store yielded nets and fix cut detection
I'm pretty sure it's actually impossible to get to the same net twice now that instructions are de-duped. I've realized that my old definition of 'same net' was wrong though - a rotated or translated net is not the same, because it maps differently onto the cuboid. So, if `prev_nets` was to be kept, it'd have to be modified to stop `shrink`ing and considering rotated nets equal.

The previous cutting detection wasn't foolproof, because it considered it okay if two squares were next to each other on the net and the cuboid but were rotated on the cuboid, which is wrong. The new logic is more robust, checking that a square on the net wouldn't be in a different spot on the surface if extruded from another existing square (so, whether an instruction is present that would put it in the same net spot but a different surface spot). That would need a cut to have it be only attached to one and not both, which we don't allow.