Allow for the same square on the net to represent different squares on the cuboids
I've done that by spinning up a separate `NetFinder` for each possible square on the cuboid that could map to (0, 0) on the bottom face of the first cuboid, each on its own thread.

The program is now able to find all 2263 nets that fold into both 1x1x5 and 1x2x3 cuboids (in ~9 mins), so I'm pretty confident that it's now exhaustive.