Don't iterate over the whole queue constantly
We did that in two places - when checking whether an instruction
already existed, and when checking if there were multiple instructions
attempting to put squares on the same spot on the net but different
spots on the surface. I was able to cover both of those by introducing
a map from net positions to all the surface positions instructions
attempt to match them with.