Parallelise ZDD construction
While this also makes it go faster, the main benefit of this is actually that it cuts down on memory usage.

When going one row at a time, you have to store the entire row of `NodeState`s at once to pass off to the next row, which is by far the biggest source of memory usage. With this new parallel approach, each row is run in parallel, and only a few thousand are queued up to be processed; the rest are discarded.
This does mean that the de-duplication is no longer 100% reliable, as we only store the hashes of `NodeState`s, which has the risk of a hash collision. Currently I'm just hoping it doesn't happen, but there are some workarounds I can think of.

I'm probably going to stop working on ZDD construction here, though, because I've now got the 1x1x7 and 1x3x3 ZDDs I was looking for. (assuming they haven't been corrupted by hash collisions, anyway.)

I also just realised I can replace all the `usize`s in `NodeState` with `u8`s and reduce memory usage by way more, but I need to go to bed so that's a problem for later.