Only spawn one thread for each core, improved UI
Since it's not like my computer could run all the rows in paralllel anyway, I decided to make it only as many rows are computed in parallel as cores you have, to further save on memory.

However, I'm mainly committing this as a reference point before trying to store `u128` sets of included/excluded edges instead of `NodeState`s.

Although I've optimised the algorithm to the point where it can fit in RAM (with the aid of memory compression), it still seems like it's some kind of bottleneck. Even though it's parallelised and should be using all 10 cores, it's only using slightly more than 1, and Activity Monitor's showing the vast majority of CPU time being used by the system (likely to do with the memory compression). So I'm going to try that technique to further reduce the memory usage and see if it helps.