Properly de-duplicate nodes
I made a critical error in my initial implementation of ZDD construction, which was that I ignored the line saying to delete the data associated with a vertex once it exited the frontier.

This meant that a lot of nodes that should have been considered equal weren't, which was causing much of my problems with memory usage, since there were a lot more nodes than there should be. It's still not light on memory and I'm not going to undo my optimisations, but it's now much much less of a problem.

So now, for the second time, I'm declaring ZDD construction done. Hopefully I'm not speaking too soon again.