Make `Net::color` stricter & fix `Net::surface_squares`
Previously, `Net::try_color` would check for whether an unfilled spot on the net led to a filled spot on the surface of the cuboid before adding it to the queue. This meant that it missed scenarios where the spot was filled in between when the position was added to the queue and when it was actually filled in, leading to invalid colourings and returning `Some` when it should be returning `None`.

I happened to discover this when attempting to get rid of the queue for an optimisation; now, `Net::try_color` performs a recursive depth-first search instead of the previous breadth-first search, so that there's no gap between a position being added to the queue and it being filled in, and all errors of this sort are caught.

I also realise that it wasn't checking for the same spot on the net leading to two different spots the surface, which it now also does.

I also fixed a mistake in `surface_squares`, which was that it assumed that you could get from any square on a face to another square with an orientation of 0 with a series of 90 degree turns, which is wrong; you can only do 180 degree turns on non-square faces, since 90 degree turns result in a different shape.

After I fixed that though, I was able to iterate over surface squares instead of net squares in `color` when picking combinations of starting net positions and face positions, which was a good bit faster.

I also optimised `Direction::turned` and `Face::adjacent_in`.