Optimise `NetFinder` some more
The most significant change here is that it no longer deals with `Pos` and `FacePos` directly; it instead uses indices, and for `FacePos` has a lookup table of which indices are in which direction from which other indices.

I also limited it to operating on 2 cuboids at once, which allows using fixed-size arrays for per-cuboid info instead of the not-particularly-fast `heapless::Vec`. Using it for more would be a pretty bad idea anyway, since it'd lead to an exponential increase in the number of threads. I think you'd be better off searching for common nets of just 2 cuboids and then filtering out the ones which also work for the third.