Delay running instructions with no follow-up
If an instruction has no follow-up instructions, we can get away with not actually running it right away. This doesn't result in any lost cases because it only results in some more instructions being able to be run later on, and the case where they don't get run will still be considered later when they're backtracked.

This saves time because we don't have to separately consider all the cases where this instruction is or isn't included.

Then, at the end, we go through all the instructions which are still valid to run, which can only be the instructions we intentionally didn't run earlier since anything else was either invalid or backtracked, and we might have lost a few that became invalid later. Then we just consider which combinations of those potential instructions yield valid nets and yield those.

Right now this is only a tiny bit faster than the status quo but I haven't optimised it much yet - that's coming in later commits.