Failed experiment: 256-way square-based ZDDs
I had the idea to try a more normal search strategy, by ignoring the net: if you do that, and instead think of solutions as mappings from squares on the surface of one net to squares on the surface of the other, you can take a more normal approach of picking a mapping for the first square, then the second, and so on, backtracking if something goes wrong. Then I thought 'Hey, I could stick this in a ZDD!'

It didn't work; the ZDD was way too big and quickly OOM-ed, even for area 22 cuboids.

The benefit of this idea, had it worked, over a normal ZDD is that it actually takes into account that there are multiple cuboids: it only produces common nets right out of the gate, whereas regular ZDDs produce _all_ nets of one cuboid, which then have to be filtered to common ones.

I might still try the non-ZDD version of this search strategy, but I don't have very high hopes for it. Thinking about it more, it has to try a _lot_ of combinations: the upper bound is 4^n * n!, for an area n cuboid, although there is some degree of detecting problems early. But I don't feel like it's early enough; it'll still have to go through loads of invalid combinations.

But I don't really have a good sense of how many combinations the normal `NetFinder` / hoarding queues algorithm goes through either, so it's kind of hard to judge. Seeing all of the node counts immediately go into the tens of millions when trying to construct a ZDD has kind of put me off this but maybe I'm just underestimating how many combinations hoarding queues have to check too.

Anyway, TL;DR I'm giving up on this idea for now but I might try it again later.