Improve logic for de-duplicating starting positions
This reduces the number of worker threads needed by different sizes by a fair bit:
- Area 22: 11 -> 9 threads.
- Area 30: 15 -> 5 threads!
- Area 46: 23 -> 18 threads.

I'm also planning to use the equivalence classes to avoid duplicating work between multiple `NetFinder`s: if one `NetFinder` comes across a mapping that's equivalent to the starting mapping of another `NetFinder`, that means that the other `NetFinder` will also be able to find its way back to the first one's starting mapping, and they'll both yield all of the nets that include both of them. So, I think I'll assign it to only be one of each of the `NetFinder`s' jobs to deal with each combination.