Use `spliter` for `EdgeIter` as well
`EdgeIter` wasn't splitting as much as it should, because it would be unable to successfully split and immediately give up. `spliter` tries again if splitting fails, which solves that and makes it properly parallelised the whole way through.

This took the time for it to find all common nets of the area 22 cuboids from 17s down to 8s.

It's still not perfect, there seems to only be one thread running in the last 2s or so, but I don't care that much since I don't think it's feasible to build a ZDD for an area-46 cuboid anyway.