Allow searching for common nets of 3 cuboids again
Switching away from fixed-length arrays has made things a bit slower (e.g. area 30 has gone from 30s to 36s) but I'm hoping that my guess that searching for common nets of 3 cuboids is faster now is right, and it'll be worth it.

I think it might be faster now because of how `NetFinder`s now skip things that are the jobs of earlier `NetFinder`s, so even though there are a crapload of them most don't have to do any work.