`equivalence_classes`: go back to only considering one cursor on one of the cuboids
Somehow I ended up believing that `equivalence_classes` covered all possible mappings, which isn't true. That led to me getting rid of the optimisation in `equivalence_classes` where on one of the cuboids, I only consider one possible starting cursor, since every net needs to have a square that maps to every cursor somewhere and I was just declaring that that's the square we're starting from.

This commit reverses that. To my surprise though, this has actually made things a bit slower: area 30's gone from 36s to 46s ish. However with the last commit trying to run `net_finder` on the area 46 cuboids would OOM so I'm doing this for the moment.

I might try and restore the unexpected speedup of covering all mappings by not actually storing every mapping in an equivalence class: instead I can store the 'canonical version' of the mappings, which will be the same for most of the mappings in an equivalence class. What that means is that you've converted all the cursors in the mapping to the version that's returned from `Cuboid::unique_cursors`. Not right now though.