Fix state-saving and improve splitting
Since we would only create the state-saving channel inside `run_finder`, a `NetFinder` wouldn't get added to our `State` until it started running, meaning that there would be loads of pending `NetFinder`s that weren't getting saved.

I also imrpoved the splitting slightly: it now aims for one more than the number of threads, to get rid of the brief gap where one of the threads is idle waiting for another to split, and made it so that `current_finders` gets incremented immediately when creating the new `NetFinder` rather than waiting till it gets received by the main thread, which stopped an explosion of like 70 `NetFinder`s appearing every time we fell below the quota.