Move most of the stuff from `lib.rs` into `geometry.rs`
This made a bunch of stuff private that was previously public, and so I had to refactor a bunch of stuff to deal with it and I think the result is generally better code.

I also noticed some problems with `NetPos` (formerly known as `IndexPos`) while trying to adopt it within `Net::color`. It doesn't properly handle going off the edge of the net: going off the left/right results in wrapping around to the next/previous row, and going off the bottom leads to overflow. Going off the top was the only case that was handled.

That's now actually documented, and I decided to lean into it by not handling going off the top either and no longer returning an `Option`.

It also handled up/down in the opposite way from `Pos`; this wasn't actually causing any problems since they aren't often mixed-and-matched, but I fixed it nonetheless so that a higher y coordinate means up everywhere, and also flipped around the order in which rows are printed to the terminal so that up on the net is actually up on the terminal.

This was kind of a waste of time tbh, but at least now the code's nicer and I realised those issues with `NetPos`.