Go back to covering all mappings with equivalence classes
The performance is now basically on-par with 8ab8f3f4b1fca985aa10592ea3180de61dfa5658, and no longer OOMs on 1x11, 1x2x7, and 1x3x5 cuboids (although it does use 2GB to start with).

I also noticed that the rayon code was quite broken: the top-level `run_finder` calls weren't actually being run in parallel, the only reason anything worked was because most of the time was being spent evaluating splits of a single `NetFinder`; that's no longer that case now that we're covering all mappings again which made me notice it.