Go back to fixing a cursor on one cuboid
It turns out that not fixing a cursor wasn't actually the reason why
area 30 got faster; in fact, doing that is actually really stupid upon
reflection.

`Combinations` will stick with the same cursor for the last cuboid until
every cursor combination for the rest of the cuboids is exhausted. Then
the next thing yielded by `unique_cursors` is all the other orientations
of the same cursor. All the other rotations of that cursor are then also
implicitly added by `SkipSet`. This means that it's now literally
impossible for any `NetFinder`s from there on to yield anything, because
any instruction that sets that square will be skipped. So you may as
well have just fixed the last `NetFinder` to that cursor!

The real reason why area 30 got faster was because the effectively-fixed
cursor happened to be a better choice than the `CursorData::new` that I
was hardcoding. Additionally, I finally figured out *why*! (i think, anyway)

It's because the average number of mappings contained in each
equivalence class is higher. This means that more mappings get skipped
in later `NetFinder`s and they run quicker.

So I've now changed `equivalence_classes` yet again to pick the set of
equivalence classes with the maximum average equivalence class size.

My job's not done yet, though - specifying 1x3x3 1x1x7 is now faster
than specifying 1x1x7 by about 2s, presumably because it's happening to
pick a set of equivalence classes that is better due to some other
factor.