Don't add invalid instructions to the queue
This idea's been in the back of my mind for a while, but I had no idea
if it would actually be a performance win; turns out it is, which is
nice!

This is accomplished by having the parents of instructions be in charge
of their associated `PosState`s rather than the instructions themselves,
which means that everything's still kept track of even if they aren't
added.

This also makes things a bit more conceptually nice, because everything
now depends solely on what instructions have actually been *run*, not
what's in the queue.

This'd be nice if I wanted to make a more general hoarding queues
abstraction, since ideally it wouldn't even communicate what's in the
queue to the struct that keeps track of state, but I think I'm going to
be jumping ship to the GPU before I get the chance. But maybe that'll
fail anyway, who knows.

As for actual numbers: area 30's gone from ~20s to ~17s, and area 34's
gone from 3m 39s to 2m 58s.