Implement `SkipSet` using a trie instead of a hash set
This isn't a slam-dunk performance win. Area 30 is a good 2s faster, but
area 34 is now ~10s slower. However, in either case the main point of
this change is for porting to the GPU, since I don't want to implement
a hash set by hand.

The weird thing about area 34 slowing down is that the new
implementation of `SkipSet` is actually significantly faster - that
isn't the problem. There seems to be some kind of weird spooky-action-
at-a-distance thing going on that causes everything else to get a bit
slower, even `backtrack` which doesn't touch `SkipSet`! Oh well.