wip 3
i wasn't checking for duplicate instructions which led to the queue
getting overrun and trashing everything else. that's fixed now but it
still isn't working.

i'm not sure if we remove potential instructions when backtracking rn...
that might be the culprit, but I think there's something else going on
(it's only running for like a couple dozen iterations, which wouldn't be
enough for say potential instructions to overrun the buffer)