fix the shader, hopefully for realsies this time
It now seems to work: the number of solutions it produces scales up
with how many iterations I limit it to. (I think it's timing out when I
put no limit on 'cause GPU APIs don't expect shaders to run this long
but I haven't put that much effort into checking.)

However the code that reads the solutions seems to be broken, but I need
to go to bed so I'll fix that later.