Remove followup instructions in reverse order
This makes sure that `InstructionSet`'s assumption that the instruction
being removed will be the last one in the list is correct. It already
worked because there can't be multiple neighbours of an instruction that
set the same square (unless it's a weird flat cuboid), but it doesn't
hurt to guarantee it more strongly.

This also improved the runtime on area 30 by ~0.05s, which is nice.
Probably something to do with cache efficiency, since the last
instruction's entry in `queued` is probably more likely to still be in
cache.