Implement `valid_checker`
This uses more LUTs than I was hoping for: 562, when it needed to use 369 or less to be able to fit 365 cores (1 per BRAM). And this isn't even the whole core (although it should be most of it). I have some ideas for how to reduce it, but I doubt I'll be able to get it under that line; oh well.