oops again again
I just can't stop making this mistake, apparently. The real LUT count for the Verilog version is 636 - basically the same as VHDL. I think the slight increase is due to `mapping_index_lookup` (formerly `skip_checker`) now having to mux between stuff.

I think I might actually be able to do this on purpose, though, and get the gains from it. What this mistake does is make it so that each neighbour is what you get by moving in a particular direction relative to the square, not the cursor like it should be. So it basically shifts `neighbours` left by the cursor's orientation. But we can actually quite easily undo that shift and get the correct result anyway. I'll give that a go momentarily.