Don't include the fixed class in mapping indices
This was the real cause of skip checking using more LUTs than before; my previous theory was nonsense, since area 46 only has one choice to check anyway and so there's no muxing.

The reason I changed it when switching to Verilog was because the size of a mapping index included all 3; I didn't realise the fixed cursor was excluded when I wrote that. And so I saw that when switching to Verilog and thought that it was included.

This had a much bigger effect than I expected - from 636 to 564 LUTs. I'm not really sure why: an individual `skip_checker` only went from 45 to 41 LUTs (they used to be 39 with VHDL).