Increase `core`'s max clock speed to 54MHz
The main optimisation I did was to move the computation of `target` and `neighbours` to the previous clock cycle. This would've accomplished nothing if they still depended on the whole critical path though, so I made them assume that the outputs of `valid_checker` were always 0 so that they could run in parallel with it.

I accomplished this my moving all the combination logic that depends on the outputs of `valid_checker` into a function, and calling it once with the correct values and a second time with the outputs set to 0.

The fact that `valid_checker`'s outputs were wrong of course meant that the computed `target` and `neighbours` was sometimes wrong, though, so I had to add an extra STALL state to switch to whenever that happened.

Now that I think about it, I didn't get quite the frequency gain out of this that I was hoping, and I wonder if it's worth the STALLs... there's still some more optimisation potential in this though, I just want to get a commit out so it doesn't feel like I'm working on rocky ground.

The other thing I changed was shrinking the number of bits in `state`; I assumed that SystemVerilog enums were special and Vivado knew that only certain values were valid even though they're theoretically 32 bits, but apparently not. I think it would've been about 50MHz without that part.