Don't use `assign` for specifying inputs
It turns out that Vivado treats unspecified inputs as inputs that are assigned the value 0, and then `assign`ing to them results in them having multiple drivers. It also seems like this might've been the cause of some weird undefined behaviour where `predicted_target_ref` ended up depending on `valid_checker` even though all the dependencies in the source code are definitely now removed.

I noticed this as part of the other main change I made, which was disabling flattening in turn with making changing `comb_main` to a proper module named `vc_dependent`. This caused Vivado to suddenly actually notice the issue and throw an error, for whatever reason.