Improve `core`'s max clock frequency to 80 MHz
As part of this, I reverted the approach used for calculating mappings in `instruction_neighbour` back to directly calculating the (unrotated) output cursors, rather than calculating and then applying offsets.

The reason I made that change in the first place was to reduce LUT count, and indeed reverting it did increase the LUT count, but the timing benefit outweighs it. (It didn't increase by nearly as much as I'd expected though, only 18 LUTs. Not sure why that is.)