Add support for multiple cores
With the more densely-packed FPGA, I had to reduce the cores' clock frequency to 60MHz to get timing to pass. Same for the system clock, although I think it's a bit lower than necessary right now: the timing improved a lot from removing DMA. But it doesn't really matter and I just want to commit this already without touching anything.