Improve splitting logic
Only ever asking the previous core to split meant that after the first core got a finder, it would split for the second core, which would split for the third, etc. This meant that by the time you got to the 80th core, it only gets one 2^-79th of a finder, which it immediately finishes and has to ask for a new one.

This commit makes it so that only the core with the lowest `base_decision` ever gets split.

This improves the runtime of area 34 from an inconsistent 18-24s to a consistent 18s.

I might just start copy-pasting the current state breakdown into my commit messages:

Time spent in each state:
  CLEAR:       0.00%
  RECEIVE:     2.34%
  RUN:        55.64%
  BACKTRACK:  25.85%
  CHECK_WAIT:  1.61%
  CHECK:       0.91%
  SOLUTION:    1.23%
  STALL:      12.37%
  SPLIT:       0.04%
  PAUSE:       0.00%