Revive GPU implementation
I was hoping that switching to the FPGA's version of skip checking would
improve things by cutting down on memory use, but seemingly not... this
takes about 7.5s for area 22, when the last mention in the commit
history says it used to take about 1.5s.

This is without any optimisation though, and I still have some ideas, so
maybe...? Either way, this implementation is at least less likely to
break since it talks via. `FinderInfo`s and doesn't require using the
same data structures as the CPU.