Split pipeline stages into separate submodules
I've done this not because I think it's structurally nicer (although
I've warmed up to it), but because it makes the schematic easier to read
and use to figure out where the LUT count is coming from.

This also fixes a bug introduced in
f05f69c080b571fae90f004ab3d1caa8d6eebb20, where we would fail to exit
`Split` state upon reaching the end of the list of decisions if the
split had already occured.

Right now I've only done this for `Core`, I might want to do
`MainPipeline` as well.